requirements.txt
bottle >= 0.12.16
ldap3 >= 2.5.2, < 3.0
pyyaml >= 3.13
voluptuous >= 0.11.5
